requirements.txt
accessible-pygments==0.0.5
alabaster==1.0.0
antlr4-python3-runtime==4.9.3
anywidget==0.9.13
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
docutils==0.21.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
huggingface-hub==0.24.6
hydra-core==1.3.2
idna==3.8
imagesize==1.4.1
iopath==0.1.10
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_bbox_widget==0.6.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
myst-parser==4.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nbsphinx==0.9.5
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
prompt_toolkit==3.0.47
psutil==6.0.0
psygnal==0.11.1
pure_eval==0.2.3
pydata-sphinx-theme==0.15.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
safetensors==0.4.4
-e git+https://github.com/facebookresearch/segment-anything-2.git@7e1596c0b6462eb1d1ba7e1492430fed95023598#egg=SAM_2
scipy==1.14.1
setuptools==70.0.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.0.2
sphinx-book-theme==1.1.3
sphinx-markdown-builder==0.6.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
supervision==0.23.0
sympy==1.12
tabulate==0.9.0
timm==1.0.9
tinycss2==1.3.0
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.9.0
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.13
